Move RITM files
