+++ web_ui/packages/smart-tools/index.ts
@@ -4,3 +4,4 @@
 export { default as OpenCVLoader } from './src/utils/opencv-loader';
 export { Watershed } from './src/watershed/watershed';
 export { type WatershedInstance, type WatershedPolygon } from './src/watershed/interfaces';
+export { RITM } from './src/ritm/ritm';
